requirements.txt
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
PyJWT==2.3.0
requests==2.31.0
scipy==1.11.3
numpy==1.26.1
scikit-learn==1.3.2
mysqlclient==2.1.1
pycountry==22.3.5
phonenumbers==8.13.25
python-barcode==0.15.1
barcode==1.0.3
pandas==2.1.3
matplotlib==3.8.1
six==1.16.0
plotly==5.18.0
requests==2.31.0

